requirements.txt
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
blis==0.7.7
bpemb==0.3.3
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
conllu==4.4.2
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.13
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
flair==0.11.3
fonttools==4.33.3
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
huggingface-hub==0.7.0
hyperopt==0.2.7
idna==3.3
importlib-metadata==3.10.1
ipykernel==6.13.1
ipython==8.4.0
Janome==0.4.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.3
jupyter-core==4.10.0
kaleido==0.2.1
kiwisolver==1.4.2
konoha==4.6.5
langcodes==3.3.0
langdetect==1.0.9
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
more-itertools==8.13.0
mpld3==0.3
murmurhash==1.0.7
nest-asyncio==1.5.5
networkx==2.8.3
nltk==3.7
numpy==1.22.4
overrides==3.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pkg_resources==0.0.0
plotly==5.8.0
pptree==3.1
preshed==3.0.6
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.1.0
regex==2022.6.2
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
segtok==1.5.11
sentencepiece==0.1.95
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlitedict==2.0.0
srsly==2.4.3
stack-data==0.2.0
tabulate==0.8.9
tenacity==8.0.1
textblob==0.17.1
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
transformers==4.19.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
Wikipedia-API==0.5.4
wrapt==1.14.1
zipp==3.8.0
